share common_ops between Expr and Matrix
